Finishing implementation of SR_EELS_Import:
The user can decide to import normal SR-EELs data sets or a calibration
measurements.
All imported files are stored at the same folder based database.
Calibration measurements are stored in sub-folders named by date and an
optional comment. Normal SR-EELS data sets are prefixed with the date.
It was necessary to add a new dependency (sc.fiji.IO_) to load dm3 files
when debugging with eclipse.