Nearly finished SR_EELS_Characterisation:
The Plugin works as expected. The plots show the correct results.

Saving results is not implemented, yet. I will switch from txt files to
using an image/stack.

The characterisation is in Java more complex than in JavaScript. JavaDoc
is necessary to understand the code. I will soon add JavaDoc.