Adding the old version of SR_EELS_CorrectionPlugin:
This version is copied from https://github.com/EFTEMj/EFTEMj.
It uses text files as input for parameters and is therefore not
compatible with SR_EELS_Characterisation. The first step was only to
import the code and include everything that is necessary to run the main
method without any error.

The next step will be to switch to the new image based data format.