Changing SR_EELS_CorrectionPlugin to use the new results.tif (part 2):
I can successfully import the new results.tif. To achieve this I had to
rewrite the internal class DataImporter.

In SR_EELS_Characterisation I had to reorder the first two columns with
fit results to match the order of columns 2 and 4.

This version does not consider the binning during the characterisation,
but the correction expects values that are scaled to match binning 1.

This correction does not correct the distortion but makes it even
stronger. This is due to a sign that is missing. The orientations of the
axes were different for the old characterisation.