Correction error in SR_EELS_CorrectionPlugin:
The orientation of the axes was not the problem that caused the
correction to make the distortion even worth.
I interchanged two axes. For the 2D polynomial fit of the borders I plot
the lateral position of the border (z values) against the position an
the energy dispersive axis (x values) and the constant y-intercept of
the border (y values). At the last commit z and y values were
interchanged.