Correcting usage of new ImagesToStack:
When removing my own implementation of ImagesToStack(), I replaced it by
``new ImagesToStack().run(null)``. This results in not showing a dialog
to set up the stack creation, resulting in "Use Titles as Labels" not
being selected.
With ``new ImagesToStack().convertImagesToStack()`` a Stack is shown and
one can select "Use Titles as Labels".