Extending EELS_SpectrumFitPlugin to support different fit functions:
New fit functions can be added by creating a new class that implements
the interface EELS_BackgroundFunction. Each of this classes, that is
stored part of the package de.m_entrup.EFTEMj_lib.lma will be detected
by using org.reflections.Reflections. This mechanism is implemented in
FunctionList.

A linear function is added to demonstrate the new feature.