Fixing a problem with stacks that don't contain loss at slice labels:
EFTEMj-ESI threw a NullPointerException when one tried to select a stack
that does not contain the energy loss at the slice label.
The commit introduces a new feature that fixes the bug and parses the
energy loss of my spectrum imaging stacks. These stacks contain a
pattern in the stack title: e.g. [418eV,748eV,30eV]. The first number is
the starting energy and the third number the energy increment. The
second number is to state the energy interval. The current
implementation supports decreasing energy losses, when the first energy
is higher than the second energy.
The user can add a custom RegExp pattern with 2 groups, where the first
group is the starting energy and the second group is the energy
increment. The ev must not be part of the group.