Adding the lateral calibration to EFTEMj-SR-EELS:
I only need to use the inverse of the max of the width function. This
value is saved at the calibration as pixel height. To get the height in
nm, one has to multiply the given value with the measured diameter of
the small filter entrance aperture. Therefore an image of the filter
entrance aperture is necessary that was recorded with the same
magnification as the SR-EELS data set.