Adding a plugin to create an EEL spectrum from an image:
This plugin is not finished, yet.
It assumes that the x axis is the energy axis.
Intensity is summed along the y axis.
A ROI can be used to limit the y region used for summation.
If there is an energy calibration in a dm3 file, it is used.
Energy calibration for other files is not implemented, yet.