Adding the energy dispersion plugins from my old EFTEMj version:
There is one plugin that is used to mange a dispersion database.
The second plugin applies an energy dispersion to an image.

The new class EnergyDispersion is used to load/save the values.
It uses the config file of EFTEMj (EFTEMj_config.xml).

The LogTool is a dependency of the imported plugins.