Adding a script to automate the SR-EELS characterization:
The script searches for all data sets in the database.
If no result for the current setting is present
the characterization of the data set is performed.

There is a bug when the characterization fails.
The threshold plugin shows a dialog the user has to confirm.
The script will pause until user interaction.