Enhancing the SR-EELS characterisation:
The py script uses only threshold methods that work well with SR-EELS data sets.
The characterisation can now use different polynomials to fit the borders.
The class Binner is used to shrink the JPEG images (instead of IJ.run...).
Adding an option to show the JPEG images during characterisation.
The limit method gets enhanced. The value is the average of mean and median.
The results class gets a new constant to define the default output types.