Fixing a bug when running the characterization on corrected images:
Corrected images can be larger than uncorrected ones.
This results in a binning of 0 due to integer division.
Finally remove outliers uses a filter size of infinity and does nothing.

The fix is to set the binning to 1 if it gets 0.
Additionally the threshold for remove outliers is calculated the same way as the limit for segmentation.